+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/instructions/ContextfulInstructionCreator.java
@@ -7,7 +7,6 @@ import info.teksol.mc.mindcode.logic.opcodes.Opcode;
 import org.jspecify.annotations.NullMarked;
 
 import java.util.List;
-import java.util.function.Consumer;
 
 import static info.teksol.mc.mindcode.logic.opcodes.Opcode.*;
 
@@ -20,12 +19,6 @@ import static info.teksol.mc.mindcode.logic.opcodes.Opcode.*;
 public interface ContextfulInstructionCreator {
 
     InstructionProcessor getProcessor();
-
-    /// Applies the provided effects to the next instruction to be created.
-    default ContextfulInstructionCreator withEffects(Consumer<LogicInstruction> decorator) {
-        getProcessor().withEffects(decorator);
-        return this;
-    }
 
     /// Allocates a new temporary variable.
     ///