+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/instructions/ContextlessInstructionCreator.java
@@ -6,7 +6,6 @@ import info.teksol.mc.mindcode.logic.opcodes.Opcode;
 import org.jspecify.annotations.NullMarked;
 
 import java.util.List;
-import java.util.function.Consumer;
 
 import static info.teksol.mc.mindcode.logic.opcodes.Opcode.*;
 
@@ -19,12 +18,6 @@ import static info.teksol.mc.mindcode.logic.opcodes.Opcode.*;
 public interface ContextlessInstructionCreator {
 
     InstructionProcessor getProcessor();
-
-    /// Applies the provided effects to the next instruction to be created.
-    default ContextlessInstructionCreator withEffects(Consumer<LogicInstruction> decorator) {
-        getProcessor().withEffects(decorator);
-        return this;
-    }
 
     LogicInstruction createInstruction(AstContext astContext, Opcode opcode, List<LogicArgument> arguments);